\name{lincongauss-package}
\alias{lincongauss-package}
\alias{lincongauss}
\docType{package}
\title{
\packageTitle{lincongauss}
}
\description{
\packageDescription{lincongauss}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lincongauss}
\packageIndices{lincongauss}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lincongauss}

Maintainer: \packageMaintainer{lincongauss}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
